% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQuery.R
\name{createQuery}
\alias{createQuery}
\title{createQuery}
\usage{
createQuery(
  theDatasource = NULL,
  theToken = NULL,
  parameters = "{}",
  description = "null",
  variables = "null",
  timefilter = "null",
  geofilter = "null",
  coordfilter = "null",
  crs = "null",
  aggregate = "null"
)
}
\arguments{
\item{theDatasource}{The id of the datasource}

\item{parameters}{Datasource parameters}

\item{description}{Optional description of this query}

\item{variables}{List of selected variables}

\item{timefilter}{Time filter}

\item{geofilter}{Spatial filter or interpolator}

\item{coordfilter}{List of additional coordinate filters}

\item{crs}{Spatial reference for filter and output}

\item{aggregate}{Aggregation operators to apply}
}
\description{
Function used to establish connection with datamesh service
}
\examples{
Query(theDatasource = "asp_5df5a74a-13d7-455a-a776-ed6c6db72644")

}
