% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConnection.R
\name{makeConnection}
\alias{makeConnection}
\title{makeConnection}
\usage{
makeConnection(token = NULL)
}
\arguments{
\item{token}{Character string with the Scopus query}
}
\description{
Function used to establish connection with datamesh service
}
\examples{
connection(token = yourToken)

}
